requirements.txt
-i https://data.pyg.org/whl/torch-1.11.0+cu113.html
torch-sparse==0.6.13
torch-scatter==2.0.9
git+https://github.com/pyg-team/pytorch_geometric.git
git+https://github.com/didriknielsen/survae_flows.git
einops
wandb